Changement de types zigzag pour la DCT (-1024, 1023)

+++ include/zigzag.h
@@ -5,11 +5,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-/* type : uint8_t
- * rtype : vecteur de uint8_t
+/* type : int16_t
+ * rtype : vecteur de int16_t
  * L'opération de réorganisation en "zig-zag" permet de représenter un
  * bloc 8 x 8 sous forme de vecteur 1D de 64 coefficients
  */
-extern void zigzag(uint8_t matrice[8][8], uint8_t tableau[64]);
+extern void zigzag(int16_t matrice[8][8], int16_t tableau[64]);
 
 #endif
